Seperate API and Provider logic
